Complete version of the game, as of the end of the game jam!

+++ World of Edge/Assets/Scripts/BulletParent.cs
@@ -36,7 +36,7 @@ public class BulletParent : MonoBehaviour {
             {
                 enemy.Flash();
             }
-            health.takeDamage((int)damage);
+            health.takeDamage(damage);
             Destroy(gameObject);
         }
         else
